Major changes to situations <--> events
Changed the way situations are created for and attached to events. This was necessary to avoid issues with overlapping events - I need to explicitly assert all situations manifesting in an Action, rather to implicitly go from the action start and end times via the concurrent States to the situations manifesting in them.